convert tests for basic pytest runs
This revealed a bit of an issue. Something has changed on the NCBI
side, and the old method of recovering coding sequences is broken
somehow. This would have been noticed if I'd have run periodic
regression tests against the remote service (but I've not been
doing). This needs to be fixed.