add --unify_seqid option
The --unify_seqid option forces the matched sequence seqID to be the same
as that of the original query sequence.

New fixtures are in tests/fixtures/targets/small_stockholm_unified and
tests/conftest.py

A new test is in tests/test_ncfp.py

Some corrections are made to the Makefile, to remove generation of deprecated
outputs.