add mechanism to resolve ambiguous UniProt GN fields
This commit updates the way UniProt is queried to grab both
the EMBL/GenBank recod number _and_ the ORF accession, so that
the ORF accession can be used in preference to the GN field
of the UniProt record.

This is necessary because of cases like that below:

>tr|A0A127QBK9|A0A127QBK9_9BURK/414-438

which has GN field:

GN=amt

This is a valid gene name, but is not the only amt gene in the
record, so ncfp was collecting the wrong record. By retriving
the ORF gene name from UniProt, the specific locus_tag for the
correct record is now recovered.