add final check to fix #46
This extends the checks of query protein sequences against GenBank
records. After the previous "final" check - trying the gene_id
field, we now also attempt to match the sequence if there is a
single CDS feature in the GenBank record. This seems to be useful
in cases where the GenBank record describes a single mRNA sequence,
and there are no other accessions to match on.